Fixed crash in CommandClass::ReadValueRefreshXML that was caused by QueryIntAttribute which takes an int pointer as argument and a uint8 pointer was provided Corrected 3 type castings that where causing a compile warning

+++ cpp/src/Node.cpp
@@ -3207,7 +3207,7 @@ void Node::ReadDeviceClasses
 			if( keyStr )
 			{
 				char* pStop;
-				uint16 key = (uint16)strtol( keyStr, &pStop, 16 );
+				uint8 key = (uint8)strtol( keyStr, &pStop, 16 );
 
 				if( !strcmp( str, "Generic" ) )
 				{
